requirements.txt
Click==7.0
cycler==0.10.0
Flask==1.0.2
Flask-Mobility==0.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
jsonpickle==1.1
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
networks==0.3.7
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
public==2019.3.22
pyparsing==2.3.1
python-dateutil==2.8.0
scipy==1.2.1
six==1.12.0
timedelta==2019.3.24
torch==1.0.1
torchvision==0.2.2.post3
Werkzeug==0.14.1
